# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontiTOGO API",
        "description": "API RESTful de la plateforme TontiTOGO — Gestion numérique des tontines au Togo avec suivi des cotisations et notifications SMS.",
        "contact": {
            "name": "Support TontiTOGO",
            "email": "support@tontitogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/responsible/tontines": {
            "get": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Liste paginée des tontines",
                "operationId": "c62b15277e10a2fd0feb26829d55e79d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des tontines"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Créer une tontine",
                "operationId": "03141c612e3e487b724dc5d61386b4fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "minimum_amount",
                                    "frequency",
                                    "start_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tontine Mensuelle"
                                    },
                                    "minimum_amount": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "example": "monthly",
                                        "enum": [
                                            "daily",
                                            "weekly",
                                            "monthly"
                                        ]
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tontine créée"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responsible/tontines/{id}": {
            "get": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Détail d'une tontine avec ses participants",
                "operationId": "acad420e5684eb9954fc484973565e02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tontine + stats + participants"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Modifier une tontine",
                "operationId": "62ae84b8ff9d721fa4cc8f35edcda1d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tontine mise à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Supprimer une tontine",
                "operationId": "5d6bb3cb5134c29c131622542f65565e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responsible/tontines/{tontine}/participants": {
            "post": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Ajouter un participant à une tontine",
                "operationId": "4b24f0c5710c5e94af6ad32f1788965a",
                "parameters": [
                    {
                        "name": "tontine",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "chosen_amount"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "chosen_amount": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "joined_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Participant ajouté"
                    },
                    "422": {
                        "description": "Erreur métier ou validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responsible/tontines/{tontine}/participants/{participant}": {
            "delete": {
                "tags": [
                    "Tontines"
                ],
                "summary": "Retirer un participant d'une tontine",
                "operationId": "81f8a182db55253589e413931281818c",
                "parameters": [
                    {
                        "name": "tontine",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "participant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Participant retiré"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}